Unify transaction mechanism — replace TransactionTemplate with @Transactional via ProcessingPersistenceService
Extracted DB operations from ProcessingServiceImpl to new ProcessingPersistenceService
with @Transactional methods. Removed TransactionTemplate, PlatformTransactionManager,
and @PostConstruct init. Race condition handling uses @Transactional(propagation = REQUIRES_NEW)
via self-injection for nested retry on DataIntegrityViolationException.
